requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
arxiv
wikipedia
langchain-text-splitters
langchain_huggingface
faiss-cpu
langchain-groq
huggingface_hub
streamlit
duckduckgo-search